requirements.txt
streamlit
langchain
transformers
langchain_community
sentence-transformers
torch
faiss-cpu
pypdf
chromadb
accelerate>=0.26.0
pdfminer.six